requirements.txt
urllib3
requests
pytest
pymysql
allure-pytest
pytest-html
pytest-assume
configparser
pyyaml